requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
docx2pdf==0.1.8
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
lxml==4.9.3
oauthlib==3.2.2
odfpy==1.4.1
Pillow==10.0.0
promise==2.3
pycparser==2.21
PyJWT==2.8.0
python-docx==0.8.11
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
watchman==0.0.1
